+++ axum-debug/tests/fail/generics.rs
@@ -1,6 +1,0 @@
-use axum_debug::debug_handler;
-
-#[debug_handler]
-async fn handler<T>() {}
-
-fn main() {}