+++ axum-debug/tests/fail/not_a_function.rs
@@ -1,6 +1,0 @@
-use axum_debug::debug_handler;
-
-#[debug_handler]
-struct A;
-
-fn main() {}