+++ axum-debug/tests/fail/not_async.rs
@@ -1,6 +1,0 @@
-use axum_debug::debug_handler;
-
-#[debug_handler]
-fn handler() {}
-
-fn main() {}