+++ axum-debug/tests/pass/mut_extractor.rs
@@ -1,9 +1,0 @@
-use axum_debug::debug_handler;
-
-#[debug_handler]
-async fn handler(mut foo: String) -> String {
-    foo += "bar";
-    foo
-}
-
-fn main() {}