+++ axum-macros/tests/debug_handler/fail/invalid_attrs.rs
@@ -1,0 +1,6 @@
+use axum_macros::debug_handler;
+
+#[debug_handler(foo)]
+async fn handler() {}
+
+fn main() {}