+++ axum-macros/tests/debug_handler/fail/not_a_function.rs
@@ -1,0 +1,6 @@
+use axum_macros::debug_handler;
+
+#[debug_handler]
+struct A;
+
+fn main() {}