+++ axum-macros/tests/debug_handler/fail/not_send.rs
@@ -1,0 +1,9 @@
+use axum_macros::debug_handler;
+
+#[debug_handler]
+async fn handler() {
+    let rc = std::rc::Rc::new(());
+    async {}.await;
+}
+
+fn main() {}