+++ axum-macros/tests/debug_handler/pass/impl_into_response.rs
@@ -1,0 +1,9 @@
+use axum_macros::debug_handler;
+use axum::response::IntoResponse;
+
+#[debug_handler]
+async fn handler() -> impl IntoResponse {
+    "hi!"
+}
+
+fn main() {}